TE-T349 Learning PHP: reduced code redundancy

+++ LoginSystem/includes/model/editproduct_model.php
@@ -1,7 +1,6 @@
 <?php
 
 declare(strict_types=1);
-
 function getCurrentProduct(object $conn, int $productId)
 {
     $query = "SELECT * FROM products where id = :id;";
@@ -11,47 +10,53 @@ function getCurrentProduct(object $conn, int $productId)
     $result =  $stmt->fetch(PDO::FETCH_ASSOC);
     return $result;
 }
-
-function getOldImagePath($conn,$productId){
+function getOldImagePath($conn, $productId)
+{
     $query = "SELECT image from products where id = :id;";
-    $stmt=$conn->prepare($query);
-    $stmt->bindParam(":id",$productId);
+    $stmt = $conn->prepare($query);
+    $stmt->bindParam(":id", $productId);
     $stmt->execute();
-    $result=$stmt->fetch(PDO::FETCH_ASSOC);
+    $result = $stmt->fetch(PDO::FETCH_ASSOC);
     return $result;
 }
-function updateWithImage(object $conn, int $productid, string $productName, string $productCategory, string $productPrice, string $productDescription, array $productImage)
-{   
-    
-    $uploadDir = "includes/uploads/";
-    $fileName = date("Y-m-d_H-i-s") . '_' . basename($productImage['name']);
-    $imagePath = $uploadDir . $fileName;
+function updateProduct(
+    object $conn,
+    int $productId,
+    string $productName,
+    string $productCategory,
+    string $productPrice,
+    string $productDescription,
+    ?array $productImage = null
+): void {
+    $fileName = null;
 
-    if (!move_uploaded_file($productImage['tmp_name'], __DIR__ . "/../../" . $imagePath)) {
-        die("Error uploading image.");
+    if ($productImage && !isFileEmpty($productImage)) {
+        $uploadDir = "includes/uploads/";
+        $fileName = date("Y-m-d_H-i-s") . '_' . basename($productImage['name']);
+        $imagePath = $uploadDir . $fileName;
+
+        if (!move_uploaded_file($productImage['tmp_name'], __DIR__ . "/../../" . $imagePath)) {
+            die("Error uploading image.");
+        }
     }
-    $query = "UPDATE products SET product_name = :productname,category_id = :categoryid, price = :price, description = :description, image = :image WHERE id = :id;";
+    $query = "UPDATE products SET 
+                product_name = :productname,
+                category_id = :categoryid,
+                price = :price,
+                description = :description";
+    if ($fileName) {
+        $query .= ", image = :image";
+    }
+    $query .= " WHERE id = :id;";
+
     $stmt = $conn->prepare($query);
     $stmt->bindParam(":productname", strtolower($productName));
     $stmt->bindParam(":categoryid", $productCategory);
     $stmt->bindParam(":price", $productPrice);
     $stmt->bindParam(":description", $productDescription);
-    $stmt->bindParam(":image", $fileName);
-    $stmt->bindParam(":id", $productid);
-    $stmt->execute();
-    
-}
-
-function updateWithoutImage(object $conn, int $productid, string $productName, string $productCategory, string $productPrice, string $productDescription)
-{
-
-    $query = "UPDATE products SET product_name = :productname,category_id = :categoryid, price = :price, description = :description WHERE id = :id;";
-    $stmt = $conn->prepare($query);
-    $stmt->bindParam(":id", $productid, PDO::PARAM_INT);
-    $stmt->bindParam(":productname", strtolower($productName));
-    $stmt->bindParam(":categoryid", $productCategory);
-    $stmt->bindParam(":price", $productPrice);
-    $stmt->bindParam(":description", $productDescription);
-
+    if ($fileName) {
+        $stmt->bindParam(":image", $fileName);
+    }
+    $stmt->bindParam(":id", $productId, PDO::PARAM_INT);
     $stmt->execute();
 }
